Modify ut0ut.cc ut_print_buf_hex()
for example, if the buf is "123", I think the output by hex should be 0x313233. If this is right, there's sth. wrong with the function.
For one byte, the left 4 bits should shift 4 bits to the right not 16.
The "&15" is not an error, but I think if we use 0x0F, the meaning is more clear!